Add default bounding box coordinates to JSON header for static meshes
With this change, all static meshes in a scene are now listed in the
JSON header along with their bounding box coordinates when the actor has
position and rotation 0. This is useful for computing oriented bounding
boxes for actors given their positions and rotations in any frame.

Other changes:
- Add field id_generated to JSON frame info. This number matches the
name of the images generated from that frame.
- Add aligned and oriented bounding box visualizer to advanced options
of ROXTracker editor details. It allows to render a small cube at the
location of bounding box vertexes of each static mesh included in a
list. There are two lists, one for render aligned bounding boxes (green
cubes) and the other for oriented bounding boxes (red cubes).
- Refactor of the name of the variable
ROXTracker::CameraStereoFocalDistances to StereoCameraBaselines; and
FROXCameraConfig::StereoFocalDistance to StereoBaseline.
- Other minor changes.