Add RenderTargets as the main way of obtaining data
With this commit, RGB, Normal and Instance Mask images are obtained
by default through RenderTargets (as Depth images) insted of taking a
high resolution screenshot from the viewport. This allows a faster
retrieving of data.
Furthermore, Instance Masks are now obtained by switching every actor
material with a plain color material, and retrieving its base color
(ignoring lighting). This method is useful for avoiding recompiling the
engine for getting the masks.

The old method for retrieving RGB, Normal and Mask images is also
available through a flag in the advanced settings of the Tracker Editor
Details.

Other changes:
- New LOG category for logs outputs made by the system (LogUnrealROX).
- Other minor changes.